Доработаны Dao классы

+++ src/main/java/com/sbt/javaschool/jdbc/homework1/dao/AbstractController.java
@@ -45,7 +45,7 @@ public abstract class AbstractController<T> {
      *
      * @param id - соответствует колонке идентификаторов в БД.
      */
-    public abstract void deleteById(long id);
+    public abstract void deleteById(int id);
 
     /**
      * Удаляет всех.
